Refactor io.* into user_graph.* (#125)
