--- src/pymatching/sparse_blossom/driver/io.cc
@@ -1,23 +1,0 @@
-// Copyright 2022 PyMatching Contributors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "pymatching/sparse_blossom/driver/io.h"
-
-#include <algorithm>
-
-double pm::merge_weights(double a, double b) {
-    auto sgn = std::copysign(1, a) * std::copysign(1, b);
-    auto signed_min = sgn * std::min(std::abs(a), std::abs(b));
-    return signed_min + std::log(1 + std::exp(-std::abs(a + b))) - std::log(1 + std::exp(-std::abs(a - b)));
-}
