--- src/pymatching/sparse_blossom/driver/io.h
@@ -1,71 +1,0 @@
-// Copyright 2022 PyMatching Contributors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef PYMATCHING2_IO_H
-#define PYMATCHING2_IO_H
-
-#include "pymatching/sparse_blossom/ints.h"
-#include "stim.h"
-
-namespace pm {
-
-const pm::weight_int NUM_DISTINCT_WEIGHTS = 1 << (sizeof(pm::weight_int) * 8 - 8);
-
-/// Computes the weight of an edge resulting from merging edges with weight `a' and weight `b', assuming each edge
-/// weight is a log-likelihood ratio log((1-p)/p) associated with the probability p of an error occurring on the
-/// edge, and that the error mechanisms associated with the two edges being merged are independent.
-///
-/// A mathematically equivalent implementation of this method would be:
-///
-///  double merge_weights(double a, double b){
-///     double p_a = 1/(1 + std::exp(a));
-///     double p_b = 1/(1 + std::exp(b));
-///     double p_both = p_a * (1 - p_b) + p_b * (1 - p_a);
-///     return std::log((1-p_both)/p_both);
-///  }
-///
-/// however this would suffer from numerical overflow issues for abs(a) >> 30 or abs(b) >> 30 which is avoided by the
-/// the implementation used here instead. See Equation (6) of https://ieeexplore.ieee.org/document/1495850 for more
-/// details.
-double merge_weights(double a, double b);
-
-template <typename Handler>
-void iter_detector_error_model_edges(
-    const stim::DetectorErrorModel &detector_error_model, const Handler &handle_dem_error) {
-    detector_error_model.iter_flatten_error_instructions([&](const stim::DemInstruction &instruction) {
-        std::vector<size_t> dets;
-        std::vector<size_t> observables;
-        double p = instruction.arg_data[0];
-        for (auto &target : instruction.target_data) {
-            if (target.is_relative_detector_id()) {
-                dets.push_back(target.val());
-            } else if (target.is_observable_id()) {
-                observables.push_back(target.val());
-            } else if (target.is_separator()) {
-                if (p > 0) {
-                    handle_dem_error(p, dets, observables);
-                    observables.clear();
-                    dets.clear();
-                }
-            }
-        }
-        if (p > 0) {
-            handle_dem_error(p, dets, observables);
-        }
-    });
-}
-
-}  // namespace pm
-
-#endif  // PYMATCHING2_IO_H
