+++ src/pymatching/sparse_blossom/driver/user_graph.cc
@@ -14,6 +14,12 @@
 
 #include "pymatching/sparse_blossom/driver/user_graph.h"
 
+double pm::merge_weights(double a, double b) {
+    auto sgn = std::copysign(1, a) * std::copysign(1, b);
+    auto signed_min = sgn * std::min(std::abs(a), std::abs(b));
+    return signed_min + std::log(1 + std::exp(-std::abs(a + b))) - std::log(1 + std::exp(-std::abs(a - b)));
+}
+
 pm::UserNode::UserNode() : is_boundary(false) {
 }
