+++ src/pymatching/sparse_blossom/driver/user_graph.h
@@ -21,13 +21,15 @@
 #include <vector>
 
 #include "pymatching/rand/rand_gen.h"
-#include "pymatching/sparse_blossom/driver/io.h"
 #include "pymatching/sparse_blossom/flooder/graph.h"
 #include "pymatching/sparse_blossom/ints.h"
 #include "pymatching/sparse_blossom/matcher/mwpm.h"
 #include "pymatching/sparse_blossom/search/search_graph.h"
+#include "stim.h"
 
 namespace pm {
+
+const pm::weight_int NUM_DISTINCT_WEIGHTS = 1 << (sizeof(pm::weight_int) * 8 - 8);
 
 struct UserEdge {
     size_t node1;
@@ -181,6 +183,50 @@ inline double UserGraph::to_matching_or_search_graph_helper(
 
 UserGraph detector_error_model_to_user_graph(
     const stim::DetectorErrorModel& detector_error_model, bool enable_correlations);
+
+/// Computes the weight of an edge resulting from merging edges with weight `a' and weight `b', assuming each edge
+/// weight is a log-likelihood ratio log((1-p)/p) associated with the probability p of an error occurring on the
+/// edge, and that the error mechanisms associated with the two edges being merged are independent.
+///
+/// A mathematically equivalent implementation of this method would be:
+///
+///  double merge_weights(double a, double b){
+///     double p_a = 1/(1 + std::exp(a));
+///     double p_b = 1/(1 + std::exp(b));
+///     double p_both = p_a * (1 - p_b) + p_b * (1 - p_a);
+///     return std::log((1-p_both)/p_both);
+///  }
+///
+/// however this would suffer from numerical overflow issues for abs(a) >> 30 or abs(b) >> 30 which is avoided by the
+/// the implementation used here instead. See Equation (6) of https://ieeexplore.ieee.org/document/1495850 for more
+/// details.
+double merge_weights(double a, double b);
+
+template <typename Handler>
+void iter_detector_error_model_edges(
+    const stim::DetectorErrorModel& detector_error_model, const Handler& handle_dem_error) {
+    detector_error_model.iter_flatten_error_instructions([&](const stim::DemInstruction& instruction) {
+        std::vector<size_t> dets;
+        std::vector<size_t> observables;
+        double p = instruction.arg_data[0];
+        for (auto& target : instruction.target_data) {
+            if (target.is_relative_detector_id()) {
+                dets.push_back(target.val());
+            } else if (target.is_observable_id()) {
+                observables.push_back(target.val());
+            } else if (target.is_separator()) {
+                if (p > 0) {
+                    handle_dem_error(p, dets, observables);
+                    observables.clear();
+                    dets.clear();
+                }
+            }
+        }
+        if (p > 0) {
+            handle_dem_error(p, dets, observables);
+        }
+    });
+}
 
 }  // namespace pm
