+++ src/pymatching/sparse_blossom/driver/user_graph.perf.cc
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "io.h"
-
 #include "mwpm_decoding.h"
 #include "pymatching/perf/util.perf.h"
